requirements.txt
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.1
python-dateutil==2.8.2
pytest==7.4.2
pytest-html==4.0.2
pyyaml==6.0.1
boto3==1.28.62